Program for mobile phone contacts.
Two classes, one holds private variables for name and cell #, the second class contains all the functions and creation of an array that utilizes instances of the Contacts class. Each method in the MobilePhone class is called in the Main by using a switch statement with the different options based on user input.